# docs/swagger.yaml
openapi: "3.0.0"
info:
  title: LAAWS Crawler Service API
  description: |
    An API for managing  crawler tasks
  contact:
    name: LOCKSS Support
    url: 'https://www.lockss.org/'
    email: lockss-support@lockss.org
  license:
    name: BSD-3-Clause
    url: 'https://www.lockss.org/support/open-source-license/'
  version: 1.0.0
servers:
  - url: 'https://laaws.lockss.org:443/'
paths:
...
  /crawls/:
    summary: Access point to all crawls
    description: Access to all crawls or for crawl creation.
    get:
      tags:
        - crawls
      summary: Get a list of active crawls.
      description: >-
        Get a list of all currently active crawls or a pageful of the list
        defined by the continuation token and size
      operationId: getCrawls
      parameters:
        - name: limit
          in: query
          description: The number of jobs per page
          required: false
          schema:
            default: 50
            type: integer
        - name: continuationToken
          in: query
          description: The continuation token of the next page of jobs to be returned.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: The requested crawls
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/jobPager'
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
    post:
      tags:
        - crawls
      summary: Request a crawl using a descriptor
      description: Use the information found in the descriptor object to initiate a crawl.
      operationId: addCrawl
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/crawlRequest'
        required: true
      responses:
        '202':
          description: The crawl request has been queued for operation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/crawlRequest'
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
    delete:
      tags:
        - crawls
      summary: Delete all of the currently queued and active crawl requests
      description: Halt and delete all of the currently queued and active crawls
      operationId: deleteCrawls
      parameters:
        - name: id
          in: query
          description: The crawl id
          required: true
          schema:
            type: string
      responses:
        '501':
          description: Not Implemented.
  /status:
    get:
      tags:
        - status
      summary: Get the status of the service
      description: Get the status of the service
      operationId: getStatus
      responses:
        '200':
          description: The status of the service
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apiStatus'
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
...
  '/crawls/{jobId}/fetched':
    summary: The list of urls fetched during the crawl.
    description: A paged in list of urls.
    get:
      tags:
      - crawls
      summary: A pagable list of fetched urls.
      description: Get a list of fetched urls.
      operationId: getCrawlFetched
      responses:
        '200':
          description: The requested fetched urls.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/urlPager'
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '409':
          description: Not Found
        '500':
          description: Internal Server Error
    parameters:
      - name: jobId
        in: path
        required: true
        schema:
          type: integer
      - name: continuationToken
        in: query
        description: '"The continuation token of the next page of jobs to be returned."'
        schema:
          type: string
      - name: limit
        in: query
        description: The number of jobs per page.
        schema:
          type: integer
  '/crawls/{jobId}/excluded':
    description: A paged in list of urls.
    get:
      tags:
      - crawls
      summary: A pagable list of excluded urls.
      description: Get a list of excluded urls.
      operationId: getCrawlExcluded
      responses:
        '200':
          description: The requested excluded urls.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/urlPager'
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '409':
          description: Not Found
        '500':
          description: Internal Server Error
    parameters:
      - name: jobId
        in: path
        description: identifier used to identify a specific crawl.
        required: true
        schema:
          type: integer
      - name: continuationToken
        in: query
        description: '"The continuation token of the next page of jobs to be returned."'
        schema:
          type: string
      - name: limit
        in: query
        description: The number of jobs per page.
        schema:
          type: integer
  '/crawls/{jobId}/notMotified':
    description: A paged in list of urls.
    get:
      tags:
      - crawls
      summary: A pagable list of notMotified urls.
      description: Get a list of notMotified urls.
      operationId: getCrawlNotModified
      responses:
        '200':
          description: The requested notMotified urls.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/urlPager'
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '409':
          description: Not Found
        '500':
          description: Internal Server Error
    parameters:
      - name: jobId
        in: path
        required: true
        schema:
          type: integer
      - name: continuationToken
        in: query
        description: '"The continuation token of the next page of jobs to be returned."'
        schema:
          type: string
      - name: limit
        in: query
        description: The number of jobs per page.
        schema:
          type: integer
  '/crawls/{jobId}/pending':
    description: A paged in list of urls.
    get:
      tags:
      - crawls
      summary: A pagable list of pending urls.
      description: Get a list of pending urls.
      operationId: getCrawlPending
      responses:
        '200':
          description: The requested modified urls.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/urlPager'
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '409':
          description: Not Found
        '500':
          description: Internal Server Error
    parameters:
      - name: jobId
        in: path
        required: true
        schema:
          type: integer
      - name: continuationToken
        in: query
        description: '"The continuation token of the next page of jobs to be returned."'
        schema:
          type: string
      - name: limit
        in: query
        description: The number of jobs per page.
        schema:
          type: integer
  '/crawls/{jobId}/parsed':
    description: A paged in list of urls.
    get:
      tags:
      - crawls
      summary: A pagable list of parsed urls.
      description: Get a list of parsed urls.
      operationId: getCrawlParsed
      responses:
        '200':
          description: The requested modified urls.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/urlPager'
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '409':
          description: Not Found
        '500':
          description: Internal Server Error
    parameters:
      - name: jobId
        in: path
        required: true
        schema:
          type: integer
      - name: continuationToken
        in: query
        description: '"The continuation token of the next page of jobs to be returned."'
        schema:
          type: string
      - name: limit
        in: query
        description: The number of jobs per page.
        schema:
          type: integer
  '/crawls/{jobId}/errored':
    description: A paged in list of urls.
    get:
      tags:
      - crawls
      summary: A pagable list of errored urls.
      description: Get a list of errored urls.
      operationId: getCrawlErrored
      responses:
        '200':
          description: The requested errored urls.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorPager'
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '409':
          description: Not Found
        '500':
          description: Internal Server Error
    parameters:
      - name: jobId
        in: path
        required: true
        schema:
          type: integer
      - name: continuationToken
        in: query
        description: '"The continuation token of the next page of jobs to be returned."'
        schema:
          type: string
      - name: limit
        in: query
        description: The number of jobs per page.
        schema:
          type: integer
  '/crawls/{jobId}':
    summary: Get information about a specific crawl.
    description: Return detailed information about a specific crawl.
    get:
      tags:
        - crawls
      summary: Get the crawl info for this job
      description: Get the job represented by this crawl id
      operationId: getCrawlById
      responses:
        '200':
          description: The crawl status of the requested crawl
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/crawlStatus'
          links:
            fetched:
              operationId: getCrawlFetched
              parameters:
                jobId: $request.path.jobId
            excluded:
              operationId: getCrawlExcluded
              parameters:
                jobId: $request.path.jobId
            notmodified:
              operationId: getCrawlNotModified
              parameters:
                jobId: $request.path.jobId
            parsed:
              operationId: getCrawlParsed
              parameters:
                jobId: $request.path.jobId
            pending:
              operationId: getCrawlPending
              parameters:
                jobId: $request.path.jobId
            errored:
              operationId: getCrawlErrored
              parameters:
                jobId: $request.path.jobId
        '401':
          description: Unauthorized
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
    delete:
      summary: Remove or stop a crawl
      tags:
      - crawls
      description: >-
        Delete a crawl given the crawl identifier, stopping any current
        processing, if necessary
      operationId: deleteCrawlById
      responses:
        '200':
          description: The deleted crawl
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/crawlRequest'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
    parameters:
      - name: jobId
        in: path
        description: identifier used to identify a specific crawl.
        required: true
        schema:
          type: integer
components:
  schemas:
    mimeCounter:
      description: A counter for mimeTypes seen during a crawl.
      required:
        - mimeType
        - counts
      type: object
      properties:
        mimeType:
          description: The mime type to count.
          type: string
        counts:
          format: int32
          description: number of urls of mimeType.
          type: integer
    pageInfo:
      description: The information related to pagination of content
      required:
        - curLink
        - continuationToken
        - resultsPerPage
        - totalCount
      type: object
      properties:
        totalCount:
          format: int32
          description: The total number of elements to be paginated
          type: integer
        resultsPerPage:
          format: int32
          description: The number of results per page
          type: integer
        continuationToken:
          description: The continuation token
          type: string
        curLink:
          description: The link to the current page
          type: string
        nextLink:
          description: The link to the next page
          type: string
    status:
      description: The existing state of a job
      required:
        - code
        - msg
      type: object
      properties:
        code:
          format: int32
          description: The numeric value for the current state
          type: integer
        msg:
          description: A text message defining the current state
          type: string
    apiStatus:
      description: The status information of the service
      required:
        - version
        - ready
      type: object
      properties:
        version:
          description: The version of the service
          type: string
        ready:
          description: The indication of whether the service is available
          type: boolean
    counter:
      description: A counter for urls.
      required:
        - count
        - counterLink
      properties:
        count:
          format: int32
          description: The number of elements
          type: integer
        counterLink:
          description: >-
            A link to the list of count elements or to a pager with count
            elements.
          type: string
    crawlRequest:
      description: A descriptor for a LOCKSS crawl.
      required:
        - auId
        - crawlKind
      type: object
      properties:
        auId:
          description: The unique au id for this crawled unit.
          type: string
        crawlKind:
          description: >-
            The kind of crawl being performed.  For now this is either new
            content or repair.
          enum:
            - newContent
            - repair
          type: string
        crawler:
          description: The crawler for this crawl
          type: string
          default: lockss
        repairList:
          description: The repair urls in a repair crawl
          type: array
          items:
            type: string
...
    crawlStatus:
      description: The status of a single crawl.
      required:
        - key
        - auId
        - auName
        - startUrls
        - type
        - startTime
        - endTime
        - status
        - isWaiting
        - isActive
        - isError
        - priority
      type: object
      properties:
        key:
          description: The id for the crawl.
          type: string
        auId:
          description: The id for the au.
          type: string
        auName:
          description: The name for the au.
          type: string
        type:
          description: The type of crawl.
          type: string
        startUrls:
          description: The array of start urls.
          type: array
          items:
            type: string
        startTime:
          format: date
          description: The time the crawl began in ISO-8601
          type: string
          example: 'yyyy-MM-ddTHH:mm:ss.SSSZ'
        endTime:
          format: date
          description: The time the crawl ended in ISO-8601
          type: string
          example: 'yyyy-MM-ddTHH:mm:ss.SSSZ'
        status:
          $ref: '#/components/schemas/status'
        isWaiting:
          description: True if the crawl wating to start.
          type: boolean
        isActive:
          description: True if the crawl is active.
          type: boolean
        isError:
          description: True if the crawl has errored.
          type: boolean
        priority:
          format: int32
          description: The priority for this crawl.
          type: integer
        fetched:
          $ref: '#/components/schemas/counter'
        excluded:
          $ref: '#/components/schemas/counter'
        notModified:
          $ref: '#/components/schemas/counter'
        parsed:
          $ref: '#/components/schemas/counter'
        sources:
          description: The sources to use for the crawl.
          type: array
          items:
            type: string
        pending:
          $ref: '#/components/schemas/counter'
        errors:
          $ref: '#/components/schemas/counter'
        mimeTypes:
          $ref: '#/components/schemas/mimeCounter'
        bytesFetched:
          format: int62
          description: The number of bytes fetched.
          type: integer
        depth:
          format: int32
          description: The depth of the crawl.
          type: integer
        refetchDepth:
          format: int32
          description: The refetch depth of the crawl.
          type: integer
        proxy:
          description: The proxy used for crawling.
          type: string
    urlPager:
      description: A Pager for urls.
      required:
        - pageInfo
        - urls
      type: object
      properties:
        pageInfo:
          $ref: '#/components/schemas/pageInfo'
        urls:
          description: An list of urls
          type: array
          items:
            type: string
    jobPager:
      description: A display page of jobs
      required:
        - jobs
        - pageInfo
      type: object
      properties:
        jobs:
          description: The jobs displayed in the page
          type: array
          items:
            $ref: '#/components/schemas/crawlStatus'
        pageInfo:
          $ref: '#/components/schemas/pageInfo'
    urlError:
      description: The detail for an error
      required:
        - url
        - errCode
        - detail
      properties:
        url:
          description: The url which caused the error.
          type: string
        errCode:
          format: int32
          description: The code for the error.
          type: integer
        detail:
          description: A detailed message related to the error.
          type: string
    errorPager:
      description: A Pager for urls.
      required:
        - pageInfo
        - errors
      type: object
      properties:
        pageInfo:
          $ref: '#/components/schemas/pageInfo'
        errors:
          description: An list of urls
          type: array
          items:
            $ref: '#/components/schemas/urlError'
  securitySchemes:
    basicAuth:
      type: http
      description: HTTP Basic Authentication. Works over `HTTP` and `HTTPS`
      scheme: basic
security:
  - basicAuth: []
...